requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0